refactor: Move Rebrand banner into its own hook
Opted to consolidate logic specific to the rebrand banner outside of the
general `useDefaultBanners` hook.
